add ability to ask SFDC for query plans
Prior to this commit, we only blindly executed queries for SFDC without
regard for how long they may take. This commit allows us to ask SFDC for
a query plan, which we can now examine and issue warnings about a table
scan if `warn_on_table_scan: true` is passed as a query option.