generate SObject modules dynamically at compile time
Prior to this commit we were using a GenServer as the basis of all the
SObject operations, which was fragile, leading to timeouts, resource
contention, etc. This commit takes a compile-time performance penalty to
hit a handful of SFDC API endpoints and dynamically generate modules for
each of the SObjects. Aside from the performance and concurrency
improvements, this should also lead to more readable code for the
library user by taking something like
`Forcex.sobject_get("Account", "1234567890ABCDEFGH")` and transforming
it, with an alias, to `Account.get("1234567890ABCDEFGH")`.