Add 'application/json' content type to json request
Prior to this commit, Salesforce would return an error
for HTTP requests with a json encoded body. The API would
complain that it did not accept the default encoding.
This commit adds the 'application/json' content type
to the request headers for all json requests, so that
the body of the request can be parsed on the other side.