Dynamically materialize sobject modules
Prior to this commit, transactional interaction
with the REST API was cumbersome, requiring the
user to provide the path along with a map of
parameters to the one of CRUD methods provided
by the forcex module. This commit adds a method,
`materialize_sobjects/1`, which creates a module
for each sobject on that particular salesforce
instance. Currently supports object creation,
deletion, object ID lookup, external ID lookup,
and object update.