Add ability to override default endpoint at login
Prior to this commit, it was not possible to provide an alternative URL
to `Forcex` when logging in.  As a result, non-production builds of
applications that depend on `Forcex` could not point to non-production
regions of Salesforce.

This commit introduces a few key changes to the client configuration and
login process:

1.  The configuration provided to the `:login` method will now be
examined for an endpoint override value.  If no endpoint is provided in
the config, it will default to `https://login.salesforce.com`

2.  If no endpoint config is provided directly, the default
configuration will examine the system environment for an endpoint
variable, as it currently does for username, password, etc.

In addition to allowing developers to provide an alternate login
endpoint during routine use of the library, it will also provide an
opportunity for the compilation process to reference an alternate
endpoint when gathering Salesforce object definitions.